requirements.txt
sqlalchemy==2.0.30
pymysql==1.1.0
python-dotenv==1.0.1
passlib==1.7.4
argon2-cffi==23.1.0
sentry-sdk==1.39.1
pytest==8.3.3
pytest-cov==5.0.0
